Rename module `signal` to `signals`
