+++ src/lib.rs
@@ -3,5 +3,5 @@ pub mod config;
 pub mod game;
 pub mod result;
 pub mod scenes;
-pub mod signal;
+pub mod signals;
 pub mod wrappers;
