+++ src/main.rs
@@ -23,6 +23,6 @@ async fn main() {
     };
     registry.init();
 
-    roketz::signal::install_signal_handler().expect("Failed to install signal handler");
+    roketz::signals::install_signal_handler().expect("Failed to install signal handler");
     roketz::game::run().await;
 }
